--- 1894.tmp
@@ -1,1 +1,0 @@
-MODOBJFILES= Otsuka_STN.o test.o
