--- 1905.tmp
@@ -1,1 +1,0 @@
-MODOBJFILES= myions.o Otsuka_STN.o
