+++ arm64/ampa.c
@@ -22,14 +22,13 @@ extern int _method3;
 extern double hoc_Exp(double);
 #endif
  
-#define nrn_init _nrn_init__AMPA_S
-#define _nrn_initial _nrn_initial__AMPA_S
-#define nrn_cur _nrn_cur__AMPA_S
-#define _nrn_current _nrn_current__AMPA_S
-#define nrn_jacob _nrn_jacob__AMPA_S
-#define nrn_state _nrn_state__AMPA_S
-#define _net_receive _net_receive__AMPA_S 
-#define release release__AMPA_S 
+#define nrn_init _nrn_init__ampa
+#define _nrn_initial _nrn_initial__ampa
+#define nrn_cur _nrn_cur__ampa
+#define _nrn_current _nrn_current__ampa
+#define nrn_jacob _nrn_jacob__ampa
+#define nrn_state _nrn_state__ampa
+#define _net_receive _net_receive__ampa 
  
 #define _threadargscomma_ _p, _ppvar, _thread, _nt,
 #define _threadargsprotocomma_ double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt,
@@ -44,27 +43,8 @@ extern double hoc_Exp(double);
  
 #define t _nt->_t
 #define dt _nt->_dt
-#define i _p[0]
-#define i_columnindex 0
-#define g _p[1]
-#define g_columnindex 1
-#define Ron _p[2]
-#define Ron_columnindex 2
-#define Roff _p[3]
-#define Roff_columnindex 3
-#define synon _p[4]
-#define synon_columnindex 4
-#define DRon _p[5]
-#define DRon_columnindex 5
-#define DRoff _p[6]
-#define DRoff_columnindex 6
-#define v _p[7]
-#define v_columnindex 7
-#define _g _p[8]
-#define _g_columnindex 8
-#define _tsav _p[9]
-#define _tsav_columnindex 9
-#define _nd_area  *_ppvar[0]._pval
+#define v _p[0]
+#define v_columnindex 0
  
 #if MAC
 #if !defined(v)
@@ -99,80 +79,31 @@ extern void hoc_reg_nmodl_text(int, const char*);
 extern void hoc_reg_nmodl_filename(int, const char*);
 #endif
 
- extern Prop* nrn_point_prop_;
- static int _pointtype;
- static void* _hoc_create_pnt(Object* _ho) { void* create_point_process(int, Object*);
- return create_point_process(_pointtype, _ho);
-}
- static void _hoc_destroy_pnt(void*);
- static double _hoc_loc_pnt(void* _vptr) {double loc_point_process(int, void*);
- return loc_point_process(_pointtype, _vptr);
-}
- static double _hoc_has_loc(void* _vptr) {double has_loc_point(void*);
- return has_loc_point(_vptr);
-}
- static double _hoc_get_loc_pnt(void* _vptr) {
- double get_loc_point_process(void*); return (get_loc_point_process(_vptr));
-}
  extern void _nrn_setdata_reg(int, void(*)(Prop*));
  static void _setdata(Prop* _prop) {
  _extcall_prop = _prop;
  }
- static void _hoc_setdata(void* _vptr) { Prop* _prop;
- _prop = ((Point_process*)_vptr)->_prop;
+ static void _hoc_setdata() {
+ Prop *_prop, *hoc_getdata_range(int);
+ _prop = hoc_getdata_range(_mechtype);
    _setdata(_prop);
- }
+ hoc_retpushx(1.);
+}
  /* connect user functions to hoc names */
  static VoidFunc hoc_intfunc[] = {
- 0,0
-};
- static Member_func _member_func[] = {
- "loc", _hoc_loc_pnt,
- "has_loc", _hoc_has_loc,
- "get_loc", _hoc_get_loc_pnt,
+ "setdata_ampa", _hoc_setdata,
  0, 0
 };
  /* declare global and static user variables */
- static int _thread1data_inuse = 0;
-static double _thread1data[2];
-#define _gth 0
-#define Alpha Alpha_AMPA_S
- double Alpha = 0.94;
-#define Beta Beta_AMPA_S
- double Beta = 0.18;
-#define Cdur Cdur_AMPA_S
- double Cdur = 0.3;
-#define Erev Erev_AMPA_S
- double Erev = 0;
-#define Rtau_AMPA_S _thread1data[0]
-#define Rtau _thread[_gth]._pval[0]
-#define Rinf_AMPA_S _thread1data[1]
-#define Rinf _thread[_gth]._pval[1]
  /* some parameters have upper and lower limits */
  static HocParmLimits _hoc_parm_limits[] = {
  0,0,0
 };
  static HocParmUnits _hoc_parm_units[] = {
- "Cdur_AMPA_S", "ms",
- "Alpha_AMPA_S", "/ms",
- "Beta_AMPA_S", "/ms",
- "Erev_AMPA_S", "mV",
- "Rtau_AMPA_S", "ms",
- "i", "nA",
- "g", "umho",
  0,0
 };
- static double Roff0 = 0;
- static double Ron0 = 0;
- static double delta_t = 0.01;
  /* connect global user variables to hoc */
  static DoubScal hoc_scdoub[] = {
- "Cdur_AMPA_S", &Cdur_AMPA_S,
- "Alpha_AMPA_S", &Alpha_AMPA_S,
- "Beta_AMPA_S", &Beta_AMPA_S,
- "Erev_AMPA_S", &Erev_AMPA_S,
- "Rinf_AMPA_S", &Rinf_AMPA_S,
- "Rtau_AMPA_S", &Rtau_AMPA_S,
  0,0
 };
  static DoubVec hoc_vdoub[] = {
@@ -182,29 +113,12 @@ static double _thread1data[2];
  static void nrn_alloc(Prop*);
 static void  nrn_init(NrnThread*, _Memb_list*, int);
 static void nrn_state(NrnThread*, _Memb_list*, int);
- static void nrn_cur(NrnThread*, _Memb_list*, int);
-static void  nrn_jacob(NrnThread*, _Memb_list*, int);
- static void _hoc_destroy_pnt(void* _vptr) {
-   destroy_point_process(_vptr);
-}
- 
-static int _ode_count(int);
-static void _ode_map(int, double**, double**, double*, Datum*, double*, int);
-static void _ode_spec(NrnThread*, _Memb_list*, int);
-static void _ode_matsol(NrnThread*, _Memb_list*, int);
- 
-#define _cvode_ieq _ppvar[3]._i
- static void _ode_matsol_instance1(_threadargsproto_);
  /* connect range variables in _p that hoc is supposed to know about */
  static const char *_mechanism[] = {
  "7.7.0",
-"AMPA_S",
+"ampa",
  0,
- "i",
- "g",
  0,
- "Ron",
- "Roff",
  0,
  0};
  
@@ -213,34 +127,13 @@ extern Prop* need_memb(Symbol*);
 static void nrn_alloc(Prop* _prop) {
 	Prop *prop_ion;
 	double *_p; Datum *_ppvar;
-  if (nrn_point_prop_) {
-	_prop->_alloc_seq = nrn_point_prop_->_alloc_seq;
-	_p = nrn_point_prop_->param;
-	_ppvar = nrn_point_prop_->dparam;
- }else{
- 	_p = nrn_prop_data_alloc(_mechtype, 10, _prop);
+ 	_p = nrn_prop_data_alloc(_mechtype, 1, _prop);
  	/*initialize range parameters*/
-  }
  	_prop->param = _p;
- 	_prop->param_size = 10;
-  if (!nrn_point_prop_) {
- 	_ppvar = nrn_prop_datum_alloc(_mechtype, 4, _prop);
-  }
- 	_prop->dparam = _ppvar;
- 	/*connect ionic variables to this model*/
+ 	_prop->param_size = 1;
  
 }
  static void _initlists();
-  /* some states have an absolute tolerance */
- static Symbol** _atollist;
- static HocStateTolerance _hoc_state_tol[] = {
- 0,0
-};
- 
-#define _tqitem &(_ppvar[2]._pvoid)
- static void _net_receive(Point_process*, double*, double);
- static void _thread_mem_init(Datum*);
- static void _thread_cleanup(Datum*);
  extern Symbol* hoc_lookup(const char*);
 extern void _nrn_thread_reg(int, int, void(*)(Datum*));
 extern void _nrn_thread_table_reg(int, void(*)(double*, Datum*, Datum*, NrnThread*, int));
@@ -250,193 +143,30 @@ extern void _cvode_abstol( Symbol**, double*, int);
  void _ampa_reg() {
 	int _vectorized = 1;
   _initlists();
- 	_pointtype = point_register_mech(_mechanism,
-	 nrn_alloc,nrn_cur, nrn_jacob, nrn_state, nrn_init,
-	 hoc_nrnpointerindex, 2,
-	 _hoc_create_pnt, _hoc_destroy_pnt, _member_func);
-  _extcall_thread = (Datum*)ecalloc(1, sizeof(Datum));
-  _thread_mem_init(_extcall_thread);
-  _thread1data_inuse = 0;
+ 	register_mech(_mechanism, nrn_alloc,(void*)0, (void*)0, (void*)0, nrn_init, hoc_nrnpointerindex, 1);
  _mechtype = nrn_get_mechtype(_mechanism[1]);
      _nrn_setdata_reg(_mechtype, _setdata);
-     _nrn_thread_reg(_mechtype, 1, _thread_mem_init);
-     _nrn_thread_reg(_mechtype, 0, _thread_cleanup);
  #if NMODL_TEXT
   hoc_reg_nmodl_text(_mechtype, nmodl_file_text);
   hoc_reg_nmodl_filename(_mechtype, nmodl_filename);
 #endif
-  hoc_register_prop_size(_mechtype, 10, 4);
-  hoc_register_dparam_semantics(_mechtype, 0, "area");
-  hoc_register_dparam_semantics(_mechtype, 1, "pntproc");
-  hoc_register_dparam_semantics(_mechtype, 2, "netsend");
-  hoc_register_dparam_semantics(_mechtype, 3, "cvodeieq");
- 	hoc_register_cvode(_mechtype, _ode_count, _ode_map, _ode_spec, _ode_matsol);
- 	hoc_register_tolerance(_mechtype, _hoc_state_tol, &_atollist);
- pnt_receive[_mechtype] = _net_receive;
- pnt_receive_size[_mechtype] = 5;
+  hoc_register_prop_size(_mechtype, 1, 0);
  	hoc_register_var(hoc_scdoub, hoc_vdoub, hoc_intfunc);
- 	ivoc_help("help ?1 AMPA_S /Users/jaberjaber/BG_Modelling/ampa.mod\n");
+ 	ivoc_help("help ?1 ampa /Users/jaberjaber/BG_Modelling/ampa.mod\n");
  hoc_register_limits(_mechtype, _hoc_parm_limits);
  hoc_register_units(_mechtype, _hoc_parm_units);
  }
 static int _reset;
-static char *modelname = "simple AMPA receptors";
+static char *modelname = "";
 
 static int error;
 static int _ninits = 0;
 static int _match_recurse=1;
 static void _modl_cleanup(){ _match_recurse=1;}
- 
-static int _ode_spec1(_threadargsproto_);
-/*static int _ode_matsol1(_threadargsproto_);*/
- static int _slist1[2], _dlist1[2];
- static int release(_threadargsproto_);
- 
-/*CVODE*/
- static int _ode_spec1 (double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt) {int _reset = 0; {
-   DRon = ( synon * Rinf - Ron ) / Rtau ;
-   DRoff = - Beta * Roff ;
-   }
- return _reset;
-}
- static int _ode_matsol1 (double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt) {
- DRon = DRon  / (1. - dt*( ( ( ( - 1.0 ) ) ) / Rtau )) ;
- DRoff = DRoff  / (1. - dt*( ( - Beta )*( 1.0 ) )) ;
-  return 0;
-}
- /*END CVODE*/
- static int release (double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt) { {
-    Ron = Ron + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / Rtau)))*(- ( ( ( ( synon )*( Rinf ) ) ) / Rtau ) / ( ( ( ( - 1.0 ) ) ) / Rtau ) - Ron) ;
-    Roff = Roff + (1. - exp(dt*(( - Beta )*( 1.0 ))))*(- ( 0.0 ) / ( ( - Beta )*( 1.0 ) ) - Roff) ;
-   }
-  return 0;
-}
- 
-static void _net_receive (Point_process* _pnt, double* _args, double _lflag) 
-{  double* _p; Datum* _ppvar; Datum* _thread; NrnThread* _nt;
-   _thread = (Datum*)0; _nt = (NrnThread*)_pnt->_vnt;   _p = _pnt->_prop->param; _ppvar = _pnt->_prop->dparam;
-  if (_tsav > t){ extern char* hoc_object_name(); hoc_execerror(hoc_object_name(_pnt->ob), ":Event arrived out of order. Must call ParallelContext.set_maxstep AFTER assigning minimum NetCon.delay");}
- _tsav = t;   if (_lflag == 1. ) {*(_tqitem) = 0;}
- {
-   if ( _lflag  == 0.0 ) {
-     _args[2] = _args[2] + 1.0 ;
-     if (  ! _args[1] ) {
-       _args[3] = _args[3] * exp ( - Beta * ( t - _args[4] ) ) ;
-       _args[4] = t ;
-       _args[1] = 1.0 ;
-       synon = synon + _args[0] ;
-         if (nrn_netrec_state_adjust && !cvode_active_){
-    /* discon state adjustment for cnexp case (rate uses no local variable) */
-    double __state = Ron;
-    double __primary = (Ron + _args[3]) - __state;
-     __primary += ( 1. - exp( 0.5*dt*( ( ( ( - 1.0 ) ) ) / Rtau ) ) )*( - ( ( ( ( synon )*( Rinf ) ) ) / Rtau ) / ( ( ( ( - 1.0 ) ) ) / Rtau ) - __primary );
-    Ron += __primary;
-  } else {
- Ron = Ron + _args[3] ;
-         }
-   if (nrn_netrec_state_adjust && !cvode_active_){
-    /* discon state adjustment for cnexp case (rate uses no local variable) */
-    double __state = Roff;
-    double __primary = (Roff - _args[3]) - __state;
-     __primary += ( 1. - exp( 0.5*dt*( ( - Beta )*( 1.0 ) ) ) )*( - ( 0.0 ) / ( ( - Beta )*( 1.0 ) ) - __primary );
-    Roff += __primary;
-  } else {
- Roff = Roff - _args[3] ;
-         }
- }
-     net_send ( _tqitem, _args, _pnt, t +  Cdur , _args[2] ) ;
-     }
-   if ( _lflag  == _args[2] ) {
-     _args[3] = _args[0] * Rinf + ( _args[3] - _args[0] * Rinf ) * exp ( - ( t - _args[4] ) / Rtau ) ;
-     _args[4] = t ;
-     synon = synon - _args[0] ;
-       if (nrn_netrec_state_adjust && !cvode_active_){
-    /* discon state adjustment for cnexp case (rate uses no local variable) */
-    double __state = Ron;
-    double __primary = (Ron - _args[3]) - __state;
-     __primary += ( 1. - exp( 0.5*dt*( ( ( ( - 1.0 ) ) ) / Rtau ) ) )*( - ( ( ( ( synon )*( Rinf ) ) ) / Rtau ) / ( ( ( ( - 1.0 ) ) ) / Rtau ) - __primary );
-    Ron += __primary;
-  } else {
- Ron = Ron - _args[3] ;
-       }
-   if (nrn_netrec_state_adjust && !cvode_active_){
-    /* discon state adjustment for cnexp case (rate uses no local variable) */
-    double __state = Roff;
-    double __primary = (Roff + _args[3]) - __state;
-     __primary += ( 1. - exp( 0.5*dt*( ( - Beta )*( 1.0 ) ) ) )*( - ( 0.0 ) / ( ( - Beta )*( 1.0 ) ) - __primary );
-    Roff += __primary;
-  } else {
- Roff = Roff + _args[3] ;
-       }
- _args[1] = 0.0 ;
-     }
-   } }
- 
-static int _ode_count(int _type){ return 2;}
- 
-static void _ode_spec(NrnThread* _nt, _Memb_list* _ml, int _type) {
-   double* _p; Datum* _ppvar; Datum* _thread;
-   Node* _nd; double _v; int _iml, _cntml;
-  _cntml = _ml->_nodecount;
-  _thread = _ml->_thread;
-  for (_iml = 0; _iml < _cntml; ++_iml) {
-    _p = _ml->_data[_iml]; _ppvar = _ml->_pdata[_iml];
-    _nd = _ml->_nodelist[_iml];
-    v = NODEV(_nd);
-     _ode_spec1 (_p, _ppvar, _thread, _nt);
- }}
- 
-static void _ode_map(int _ieq, double** _pv, double** _pvdot, double* _pp, Datum* _ppd, double* _atol, int _type) { 
-	double* _p; Datum* _ppvar;
- 	int _i; _p = _pp; _ppvar = _ppd;
-	_cvode_ieq = _ieq;
-	for (_i=0; _i < 2; ++_i) {
-		_pv[_i] = _pp + _slist1[_i];  _pvdot[_i] = _pp + _dlist1[_i];
-		_cvode_abstol(_atollist, _atol, _i);
-	}
- }
- 
-static void _ode_matsol_instance1(_threadargsproto_) {
- _ode_matsol1 (_p, _ppvar, _thread, _nt);
- }
- 
-static void _ode_matsol(NrnThread* _nt, _Memb_list* _ml, int _type) {
-   double* _p; Datum* _ppvar; Datum* _thread;
-   Node* _nd; double _v; int _iml, _cntml;
-  _cntml = _ml->_nodecount;
-  _thread = _ml->_thread;
-  for (_iml = 0; _iml < _cntml; ++_iml) {
-    _p = _ml->_data[_iml]; _ppvar = _ml->_pdata[_iml];
-    _nd = _ml->_nodelist[_iml];
-    v = NODEV(_nd);
- _ode_matsol_instance1(_threadargs_);
- }}
- 
-static void _thread_mem_init(Datum* _thread) {
-  if (_thread1data_inuse) {_thread[_gth]._pval = (double*)ecalloc(2, sizeof(double));
- }else{
- _thread[_gth]._pval = _thread1data; _thread1data_inuse = 1;
- }
- }
- 
-static void _thread_cleanup(Datum* _thread) {
-  if (_thread[_gth]._pval == _thread1data) {
-   _thread1data_inuse = 0;
-  }else{
-   free((void*)_thread[_gth]._pval);
-  }
- }
 
 static void initmodel(double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt) {
   int _i; double _save;{
-  Roff = Roff0;
-  Ron = Ron0;
- {
-   Rinf = Alpha / ( Alpha + Beta ) ;
-   Rtau = 1.0 / ( Alpha + Beta ) ;
-   synon = 0.0 ;
-   }
- 
+
 }
 }
 
@@ -450,7 +180,6 @@ _cntml = _ml->_nodecount;
 _thread = _ml->_thread;
 for (_iml = 0; _iml < _cntml; ++_iml) {
  _p = _ml->_data[_iml]; _ppvar = _ml->_pdata[_iml];
- _tsav = -1e20;
 #if CACHEVEC
   if (use_cachevec) {
     _v = VEC_V(_ni[_iml]);
@@ -465,75 +194,8 @@ for (_iml = 0; _iml < _cntml; ++_iml) {
 }
 }
 
-static double _nrn_current(double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt, double _v){double _current=0.;v=_v;{ {
-   g = ( Ron + Roff ) * 1.0 ;
-   i = g * ( v - Erev ) ;
-   }
- _current += i;
-
+static double _nrn_current(double* _p, Datum* _ppvar, Datum* _thread, NrnThread* _nt, double _v){double _current=0.;v=_v;{
 } return _current;
-}
-
-static void nrn_cur(NrnThread* _nt, _Memb_list* _ml, int _type) {
-double* _p; Datum* _ppvar; Datum* _thread;
-Node *_nd; int* _ni; double _rhs, _v; int _iml, _cntml;
-#if CACHEVEC
-    _ni = _ml->_nodeindices;
-#endif
-_cntml = _ml->_nodecount;
-_thread = _ml->_thread;
-for (_iml = 0; _iml < _cntml; ++_iml) {
- _p = _ml->_data[_iml]; _ppvar = _ml->_pdata[_iml];
-#if CACHEVEC
-  if (use_cachevec) {
-    _v = VEC_V(_ni[_iml]);
-  }else
-#endif
-  {
-    _nd = _ml->_nodelist[_iml];
-    _v = NODEV(_nd);
-  }
- _g = _nrn_current(_p, _ppvar, _thread, _nt, _v + .001);
- 	{ _rhs = _nrn_current(_p, _ppvar, _thread, _nt, _v);
- 	}
- _g = (_g - _rhs)/.001;
- _g *=  1.e2/(_nd_area);
- _rhs *= 1.e2/(_nd_area);
-#if CACHEVEC
-  if (use_cachevec) {
-	VEC_RHS(_ni[_iml]) -= _rhs;
-  }else
-#endif
-  {
-	NODERHS(_nd) -= _rhs;
-  }
- 
-}
- 
-}
-
-static void nrn_jacob(NrnThread* _nt, _Memb_list* _ml, int _type) {
-double* _p; Datum* _ppvar; Datum* _thread;
-Node *_nd; int* _ni; int _iml, _cntml;
-#if CACHEVEC
-    _ni = _ml->_nodeindices;
-#endif
-_cntml = _ml->_nodecount;
-_thread = _ml->_thread;
-for (_iml = 0; _iml < _cntml; ++_iml) {
- _p = _ml->_data[_iml];
-#if CACHEVEC
-  if (use_cachevec) {
-	VEC_D(_ni[_iml]) += _g;
-  }else
-#endif
-  {
-     _nd = _ml->_nodelist[_iml];
-	NODED(_nd) += _g;
-  }
- 
-}
- 
 }
 
 static void nrn_state(NrnThread* _nt, _Memb_list* _ml, int _type) {
@@ -558,8 +220,7 @@ for (_iml = 0; _iml < _cntml; ++_iml) {
   }
  v=_v;
 {
- {   release(_p, _ppvar, _thread, _nt);
-  }}}
+}}
 
 }
 
@@ -569,8 +230,6 @@ static void _initlists(){
  double _x; double* _p = &_x;
  int _i; static int _first = 1;
   if (!_first) return;
- _slist1[0] = Ron_columnindex;  _dlist1[0] = DRon_columnindex;
- _slist1[1] = Roff_columnindex;  _dlist1[1] = DRoff_columnindex;
 _first = 0;
 }
 
@@ -581,136 +240,5 @@ _first = 0;
 #if NMODL_TEXT
 static const char* nmodl_filename = "/Users/jaberjaber/BG_Modelling/ampa.mod";
 static const char* nmodl_file_text = 
-  "TITLE simple AMPA receptors\n"
-  "\n"
-  "COMMENT\n"
-  "-----------------------------------------------------------------------------\n"
-  "\n"
-  "	Simple model for glutamate AMPA receptors\n"
-  "	=========================================\n"
-  "\n"
-  "  - FIRST-ORDER KINETICS, FIT TO WHOLE-CELL RECORDINGS\n"
-  "\n"
-  "    Whole-cell recorded postsynaptic currents mediated by AMPA/Kainate\n"
-  "    receptors (Xiang et al., J. Neurophysiol. 71: 2552-2556, 1994) were used\n"
-  "    to estimate the parameters of the present model; the fit was performed\n"
-  "    using a simplex algorithm (see Destexhe et al., J. Computational Neurosci.\n"
-  "    1: 195-230, 1994).\n"
-  "\n"
-  "  - SHORT PULSES OF TRANSMITTER (0.3 ms, 0.5 mM)\n"
-  "\n"
-  "    The simplified model was obtained from a detailed synaptic model that \n"
-  "    included the release of transmitter in adjacent terminals, its lateral \n"
-  "    diffusion and uptake, and its binding on postsynaptic receptors (Destexhe\n"
-  "    and Sejnowski, 1995).  Short pulses of transmitter with first-order\n"
-  "    kinetics were found to be the best fast alternative to represent the more\n"
-  "    detailed models.\n"
-  "\n"
-  "  - ANALYTIC EXPRESSION\n"
-  "\n"
-  "    The first-order model can be solved analytically, leading to a very fast\n"
-  "    mechanism for simulating synapses, since no differential equation must be\n"
-  "    solved (see references below).\n"
-  "\n"
-  "\n"
-  "\n"
-  "References\n"
-  "\n"
-  "   Destexhe, A., Mainen, Z.F. and Sejnowski, T.J.  An efficient method for\n"
-  "   computing synaptic conductances based on a kinetic model of receptor binding\n"
-  "   Neural Computation 6: 10-14, 1994.  \n"
-  "\n"
-  "   Destexhe, A., Mainen, Z.F. and Sejnowski, T.J. Synthesis of models for\n"
-  "   excitable membranes, synaptic transmission and neuromodulation using a \n"
-  "   common kinetic formalism, Journal of Computational Neuroscience 1: \n"
-  "   195-230, 1994.\n"
-  "\n"
-  "-----------------------------------------------------------------------------\n"
-  "ENDCOMMENT\n"
-  "\n"
-  "\n"
-  "\n"
-  "NEURON {\n"
-  "	THREADSAFE\n"
-  "	POINT_PROCESS AMPA_S\n"
-  "	RANGE R, g\n"
-  "	NONSPECIFIC_CURRENT i\n"
-  "	GLOBAL Cdur, Alpha, Beta, Erev, Rinf, Rtau\n"
-  "}\n"
-  "UNITS {\n"
-  "	(nA) = (nanoamp)\n"
-  "	(mV) = (millivolt)\n"
-  "	(umho) = (micromho)\n"
-  "	(mM) = (milli/liter)\n"
-  "}\n"
-  "\n"
-  "PARAMETER {\n"
-  "\n"
-  "	Cdur	= 0.3	(ms)		: transmitter duration (rising phase)\n"
-  "	Alpha	= 0.94	(/ms)	: forward (binding) rate\n"
-  "	Beta	= 0.18	(/ms)		: backward (unbinding) rate\n"
-  "	Erev	= 0	(mV)		: reversal potential\n"
-  "}\n"
-  "\n"
-  "\n"
-  "ASSIGNED {\n"
-  "	v		(mV)		: postsynaptic voltage\n"
-  "	i 		(nA)		: current = g*(v - Erev)\n"
-  "	g 		(umho)		: conductance\n"
-  "	Rinf				: steady state channels open\n"
-  "	Rtau		(ms)		: time constant of channel binding\n"
-  "	synon\n"
-  "}\n"
-  "\n"
-  "STATE {Ron Roff}\n"
-  "\n"
-  "INITIAL {\n"
-  "	Rinf = Alpha / (Alpha + Beta)\n"
-  "	Rtau = 1 / (Alpha + Beta)\n"
-  "	synon = 0\n"
-  "}\n"
-  "\n"
-  "BREAKPOINT {\n"
-  "	SOLVE release METHOD cnexp\n"
-  "	g = (Ron + Roff)*1(umho)\n"
-  "	i = g*(v - Erev)\n"
-  "}\n"
-  "\n"
-  "DERIVATIVE release {\n"
-  "	Ron' = (synon*Rinf - Ron)/Rtau\n"
-  "	Roff' = -Beta*Roff\n"
-  "}\n"
-  "\n"
-  ": following supports both saturation from single input and\n"
-  ": summation from multiple inputs\n"
-  ": if spike occurs during CDur then new off time is t + CDur\n"
-  ": ie. transmitter concatenates but does not summate\n"
-  ": Note: automatic initialization of all reference args to 0 except first\n"
-  "\n"
-  "NET_RECEIVE(weight, on, nspike, r0, t0 (ms)) {\n"
-  "	: flag is an implicit argument of NET_RECEIVE and  normally 0\n"
-  "        if (flag == 0) { : a spike, so turn on if not already in a Cdur pulse\n"
-  "		nspike = nspike + 1\n"
-  "		if (!on) {\n"
-  "			r0 = r0*exp(-Beta*(t - t0))\n"
-  "			t0 = t\n"
-  "			on = 1\n"
-  "			synon = synon + weight\n"
-  "			Ron = Ron + r0\n"
-  "			Roff = Roff - r0\n"
-  "		}\n"
-  "		: come again in Cdur with flag = current value of nspike\n"
-  "		net_send(Cdur, nspike)\n"
-  "        }\n"
-  "	if (flag == nspike) { : if this associated with last spike then turn off\n"
-  "		r0 = weight*Rinf + (r0 - weight*Rinf)*exp(-(t - t0)/Rtau)\n"
-  "		t0 = t\n"
-  "		synon = synon - weight\n"
-  "		Ron = Ron - r0\n"
-  "		Roff = Roff + r0\n"
-  "		on = 0\n"
-  "	}\n"
-  "}\n"
-  "\n"
   ;
 #endif
