+++ arm64/mod_func.cpp
@@ -7,19 +7,16 @@ extern "C" {
 #endif
 
 extern void _Otsuka_STN_reg(void);
-extern void _ampa_reg(void);
 extern void _myions_reg(void);
 
 void modl_reg() {
   if (!nrn_nobanner_) if (nrnmpi_myid < 1) {
     fprintf(stderr, "Additional mechanisms from files\n");
     fprintf(stderr, " \"Otsuka_STN.mod\"");
-    fprintf(stderr, " \"ampa.mod\"");
     fprintf(stderr, " \"myions.mod\"");
     fprintf(stderr, "\n");
   }
   _Otsuka_STN_reg();
-  _ampa_reg();
   _myions_reg();
 }
